LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-outside-circle-002-expected.txt


FAIL circle(at 50%) serializes as circle(at 50% 50%) - inline assert_equals: expected "circle(at 50% 50%)" but got "circle(at 50% center)"
FAIL circle(at 50px) serializes as circle(at 50px 50%) - inline assert_equals: expected "circle(at 50px 50%)" but got "circle(at 50px center)"
PASS circle(at 50% 50%) serializes as circle(at 50% 50%) - inline
PASS circle(at 50% 50px) serializes as circle(at 50% 50px) - inline
PASS circle(at 50px 50%) serializes as circle(at 50px 50%) - inline
PASS circle(at 50px 50px) serializes as circle(at 50px 50px) - inline
FAIL circle(at left) serializes as circle(at 0% 50%) - inline assert_equals: expected "circle(at 0% 50%)" but got "circle(at left center)"
FAIL circle(at top) serializes as circle(at 50% 0%) - inline assert_equals: expected "circle(at 50% 0%)" but got "circle(at center top)"
FAIL circle(at right) serializes as circle(at 100% 50%) - inline assert_equals: expected "circle(at 100% 50%)" but got "circle(at right center)"
FAIL circle(at bottom) serializes as circle(at 50% 100%) - inline assert_equals: expected "circle(at 50% 100%)" but got "circle(at center bottom)"
FAIL circle(at center) serializes as circle(at 50% 50%) - inline assert_equals: expected "circle(at 50% 50%)" but got "circle(at center center)"
FAIL circle(at left top) serializes as circle(at 0% 0%) - inline assert_equals: expected "circle(at 0% 0%)" but got "circle(at left top)"
FAIL circle(at left bottom) serializes as circle(at 0% 100%) - inline assert_equals: expected "circle(at 0% 100%)" but got "circle(at left bottom)"
FAIL circle(at left center) serializes as circle(at 0% 50%) - inline assert_equals: expected "circle(at 0% 50%)" but got "circle(at left center)"
FAIL circle(at top left) serializes as circle(at 0% 0%) - inline assert_equals: expected "circle(at 0% 0%)" but got "circle(at left top)"
FAIL circle(at top right) serializes as circle(at 100% 0%) - inline assert_equals: expected "circle(at 100% 0%)" but got "circle(at right top)"
FAIL circle(at top center) serializes as circle(at 50% 0%) - inline assert_equals: expected "circle(at 50% 0%)" but got "circle(at center top)"
FAIL circle(at right top) serializes as circle(at 100% 0%) - inline assert_equals: expected "circle(at 100% 0%)" but got "circle(at right top)"
FAIL circle(at right bottom) serializes as circle(at 100% 100%) - inline assert_equals: expected "circle(at 100% 100%)" but got "circle(at right bottom)"
FAIL circle(at right center) serializes as circle(at 100% 50%) - inline assert_equals: expected "circle(at 100% 50%)" but got "circle(at right center)"
FAIL circle(at bottom left) serializes as circle(at 0% 100%) - inline assert_equals: expected "circle(at 0% 100%)" but got "circle(at left bottom)"
FAIL circle(at bottom right) serializes as circle(at 100% 100%) - inline assert_equals: expected "circle(at 100% 100%)" but got "circle(at right bottom)"
FAIL circle(at bottom center) serializes as circle(at 50% 100%) - inline assert_equals: expected "circle(at 50% 100%)" but got "circle(at center bottom)"
FAIL circle(at center top) serializes as circle(at 50% 0%) - inline assert_equals: expected "circle(at 50% 0%)" but got "circle(at center top)"
FAIL circle(at center left) serializes as circle(at 0% 50%) - inline assert_equals: expected "circle(at 0% 50%)" but got "circle(at left center)"
FAIL circle(at center right) serializes as circle(at 100% 50%) - inline assert_equals: expected "circle(at 100% 50%)" but got "circle(at right center)"
FAIL circle(at center bottom) serializes as circle(at 50% 100%) - inline assert_equals: expected "circle(at 50% 100%)" but got "circle(at center bottom)"
FAIL circle(at center center) serializes as circle(at 50% 50%) - inline assert_equals: expected "circle(at 50% 50%)" but got "circle(at center center)"
FAIL circle(at left 50%) serializes as circle(at 0% 50%) - inline assert_equals: expected "circle(at 0% 50%)" but got "circle(at left 50%)"
FAIL circle(at left 50px) serializes as circle(at 0% 50px) - inline assert_equals: expected "circle(at 0% 50px)" but got "circle(at left 50px)"
FAIL circle(at 50% top) serializes as circle(at 50% 0%) - inline assert_equals: expected "circle(at 50% 0%)" but got "circle(at 50% top)"
FAIL circle(at 50px top) serializes as circle(at 50px 0%) - inline assert_equals: expected "circle(at 50px 0%)" but got "circle(at 50px top)"
FAIL circle(at right 80%) serializes as circle(at 100% 80%) - inline assert_equals: expected "circle(at 100% 80%)" but got "circle(at right 80%)"
FAIL circle(at right 80px) serializes as circle(at 100% 80px) - inline assert_equals: expected "circle(at 100% 80px)" but got "circle(at right 80px)"
FAIL circle(at 70% bottom) serializes as circle(at 70% 100%) - inline assert_equals: expected "circle(at 70% 100%)" but got "circle(at 70% bottom)"
FAIL circle(at 70px bottom) serializes as circle(at 70px 100%) - inline assert_equals: expected "circle(at 70px 100%)" but got "circle(at 70px bottom)"
FAIL circle(at center 60%) serializes as circle(at 50% 60%) - inline assert_equals: expected "circle(at 50% 60%)" but got "circle(at center 60%)"
FAIL circle(at center 60px) serializes as circle(at 50% 60px) - inline assert_equals: expected "circle(at 50% 60px)" but got "circle(at center 60px)"
FAIL circle(at 60% center) serializes as circle(at 60% 50%) - inline assert_equals: expected "circle(at 60% 50%)" but got "circle(at 60% center)"
FAIL circle(at 60px center) serializes as circle(at 60px 50%) - inline assert_equals: expected "circle(at 60px 50%)" but got "circle(at 60px center)"
FAIL circle(at left 50% top 50%) serializes as circle(at 50% 50%) - inline assert_equals: expected "circle(at 50% 50%)" but got "circle(at left 50% top 50%)"
FAIL circle(at left 50% top 50px) serializes as circle(at 50% 50px) - inline assert_equals: expected "circle(at 50% 50px)" but got "circle(at left 50% top 50px)"
FAIL circle(at left 50% bottom 70%) serializes as circle(at 50% calc(30%)) - inline assert_equals: expected "circle(at 50% calc(30%))" but got "circle(at left 50% bottom 70%)"
FAIL circle(at left 50% bottom 70px) serializes as circle(at 50% calc(100% - 70px)) - inline assert_equals: expected "circle(at 50% calc(100% - 70px))" but got "circle(at left 50% bottom 70px)"
FAIL circle(at left 50px top 50%) serializes as circle(at 50px 50%) - inline assert_equals: expected "circle(at 50px 50%)" but got "circle(at left 50px top 50%)"
FAIL circle(at left 50px top 50px) serializes as circle(at 50px 50px) - inline assert_equals: expected "circle(at 50px 50px)" but got "circle(at left 50px top 50px)"
FAIL circle(at left 50px bottom 70%) serializes as circle(at 50px calc(30%)) - inline assert_equals: expected "circle(at 50px calc(30%))" but got "circle(at left 50px bottom 70%)"
FAIL circle(at top 50% left 50%) serializes as circle(at 50% 50%) - inline assert_equals: expected "circle(at 50% 50%)" but got "circle(at left 50% top 50%)"
FAIL circle(at top 50% left 50px) serializes as circle(at 50px 50%) - inline assert_equals: expected "circle(at 50px 50%)" but got "circle(at left 50px top 50%)"
FAIL circle(at top 50% right 80%) serializes as circle(at calc(20%) 50%) - inline assert_equals: expected "circle(at calc(20%) 50%)" but got "circle(at right 80% top 50%)"
FAIL circle(at top 50% right 80px) serializes as circle(at calc(100% - 80px) 50%) - inline assert_equals: expected "circle(at calc(100% - 80px) 50%)" but got "circle(at right 80px top 50%)"
FAIL circle(at top 50px left 50%) serializes as circle(at 50% 50px) - inline assert_equals: expected "circle(at 50% 50px)" but got "circle(at left 50% top 50px)"
FAIL circle(at top 50px left 50px) serializes as circle(at 50px 50px) - inline assert_equals: expected "circle(at 50px 50px)" but got "circle(at left 50px top 50px)"
FAIL circle(at top 50px right 80%) serializes as circle(at calc(20%) 50px) - inline assert_equals: expected "circle(at calc(20%) 50px)" but got "circle(at right 80% top 50px)"
FAIL circle(at bottom 70% left 50%) serializes as circle(at 50% calc(30%)) - inline assert_equals: expected "circle(at 50% calc(30%))" but got "circle(at left 50% bottom 70%)"
FAIL circle(at bottom 70% left 50px) serializes as circle(at 50px calc(30%)) - inline assert_equals: expected "circle(at 50px calc(30%))" but got "circle(at left 50px bottom 70%)"
FAIL circle(at bottom 70% right 80%) serializes as circle(at calc(20%) calc(30%)) - inline assert_equals: expected "circle(at calc(20%) calc(30%))" but got "circle(at right 80% bottom 70%)"
FAIL circle(at bottom 70% right 80px) serializes as circle(at calc(100% - 80px) calc(30%)) - inline assert_equals: expected "circle(at calc(100% - 80px) calc(30%))" but got "circle(at right 80px bottom 70%)"
FAIL circle(at bottom 70px left 50%) serializes as circle(at 50% calc(100% - 70px)) - inline assert_equals: expected "circle(at 50% calc(100% - 70px))" but got "circle(at left 50% bottom 70px)"
FAIL circle(at bottom 70px right 50%) serializes as circle(at calc(50%) calc(100% - 70px)) - inline assert_equals: expected "circle(at calc(50%) calc(100% - 70px))" but got "circle(at right 50% bottom 70px)"
FAIL circle(at bottom 70px right 80px) serializes as circle(at calc(100% - 80px) calc(100% - 70px)) - inline assert_equals: expected "circle(at calc(100% - 80px) calc(100% - 70px))" but got "circle(at right 80px bottom 70px)"
FAIL circle(at right 80% top 50%) serializes as circle(at calc(20%) 50%) - inline assert_equals: expected "circle(at calc(20%) 50%)" but got "circle(at right 80% top 50%)"
FAIL circle(at right 80% top 50px) serializes as circle(at calc(20%) 50px) - inline assert_equals: expected "circle(at calc(20%) 50px)" but got "circle(at right 80% top 50px)"
FAIL circle(at right 80% bottom 70%) serializes as circle(at calc(20%) calc(30%)) - inline assert_equals: expected "circle(at calc(20%) calc(30%))" but got "circle(at right 80% bottom 70%)"
FAIL circle(at right 80% bottom 70px) serializes as circle(at calc(20%) calc(100% - 70px)) - inline assert_equals: expected "circle(at calc(20%) calc(100% - 70px))" but got "circle(at right 80% bottom 70px)"
FAIL circle(at right 80px top 50%) serializes as circle(at calc(100% - 80px) 50%) - inline assert_equals: expected "circle(at calc(100% - 80px) 50%)" but got "circle(at right 80px top 50%)"
FAIL circle(at right 80px bottom 70%) serializes as circle(at calc(100% - 80px) calc(30%)) - inline assert_equals: expected "circle(at calc(100% - 80px) calc(30%))" but got "circle(at right 80px bottom 70%)"
FAIL circle(at right 80px bottom 70px) serializes as circle(at calc(100% - 80px) calc(100% - 70px)) - inline assert_equals: expected "circle(at calc(100% - 80px) calc(100% - 70px))" but got "circle(at right 80px bottom 70px)"
PASS circle(at 50%) serializes as circle(at 50% 50%) - computed
PASS circle(at 50px) serializes as circle(at 50px 50%) - computed
PASS circle(at 50% 50%) serializes as circle(at 50% 50%) - computed
PASS circle(at 50% 50px) serializes as circle(at 50% 50px) - computed
PASS circle(at 50px 50%) serializes as circle(at 50px 50%) - computed
PASS circle(at 50px 50px) serializes as circle(at 50px 50px) - computed
PASS circle(at left) serializes as circle(at 0% 50%) - computed
PASS circle(at top) serializes as circle(at 50% 0%) - computed
PASS circle(at right) serializes as circle(at 100% 50%) - computed
PASS circle(at bottom) serializes as circle(at 50% 100%) - computed
PASS circle(at center) serializes as circle(at 50% 50%) - computed
PASS circle(at left top) serializes as circle(at 0% 0%) - computed
PASS circle(at left bottom) serializes as circle(at 0% 100%) - computed
PASS circle(at left center) serializes as circle(at 0% 50%) - computed
PASS circle(at top left) serializes as circle(at 0% 0%) - computed
PASS circle(at top right) serializes as circle(at 100% 0%) - computed
PASS circle(at top center) serializes as circle(at 50% 0%) - computed
PASS circle(at right top) serializes as circle(at 100% 0%) - computed
PASS circle(at right bottom) serializes as circle(at 100% 100%) - computed
PASS circle(at right center) serializes as circle(at 100% 50%) - computed
PASS circle(at bottom left) serializes as circle(at 0% 100%) - computed
PASS circle(at bottom right) serializes as circle(at 100% 100%) - computed
PASS circle(at bottom center) serializes as circle(at 50% 100%) - computed
PASS circle(at center top) serializes as circle(at 50% 0%) - computed
PASS circle(at center left) serializes as circle(at 0% 50%) - computed
PASS circle(at center right) serializes as circle(at 100% 50%) - computed
PASS circle(at center bottom) serializes as circle(at 50% 100%) - computed
PASS circle(at center center) serializes as circle(at 50% 50%) - computed
PASS circle(at left 50%) serializes as circle(at 0% 50%) - computed
PASS circle(at left 50px) serializes as circle(at 0% 50px) - computed
PASS circle(at 50% top) serializes as circle(at 50% 0%) - computed
PASS circle(at 50px top) serializes as circle(at 50px 0%) - computed
PASS circle(at right 80%) serializes as circle(at 100% 80%) - computed
PASS circle(at right 80px) serializes as circle(at 100% 80px) - computed
PASS circle(at 70% bottom) serializes as circle(at 70% 100%) - computed
PASS circle(at 70px bottom) serializes as circle(at 70px 100%) - computed
PASS circle(at center 60%) serializes as circle(at 50% 60%) - computed
PASS circle(at center 60px) serializes as circle(at 50% 60px) - computed
PASS circle(at 60% center) serializes as circle(at 60% 50%) - computed
PASS circle(at 60px center) serializes as circle(at 60px 50%) - computed
PASS circle(at left 50% top 50%) serializes as circle(at 50% 50%) - computed
PASS circle(at left 50% top 50px) serializes as circle(at 50% 50px) - computed
PASS circle(at left 50% bottom 70%) serializes as circle(at 50% 30%) - computed
PASS circle(at left 50% bottom 70px) serializes as circle(at 50% calc(100% - 70px)) - computed
PASS circle(at left 50px top 50%) serializes as circle(at 50px 50%) - computed
PASS circle(at left 50px top 50px) serializes as circle(at 50px 50px) - computed
PASS circle(at left 50px bottom 70%) serializes as circle(at 50px 30%) - computed
PASS circle(at top 50% left 50%) serializes as circle(at 50% 50%) - computed
PASS circle(at top 50% left 50px) serializes as circle(at 50px 50%) - computed
PASS circle(at top 50% right 80%) serializes as circle(at 20% 50%) - computed
PASS circle(at top 50% right 80px) serializes as circle(at calc(100% - 80px) 50%) - computed
PASS circle(at top 50px left 50%) serializes as circle(at 50% 50px) - computed
PASS circle(at top 50px left 50px) serializes as circle(at 50px 50px) - computed
PASS circle(at top 50px right 80%) serializes as circle(at 20% 50px) - computed
PASS circle(at bottom 70% left 50%) serializes as circle(at 50% 30%) - computed
PASS circle(at bottom 70% left 50px) serializes as circle(at 50px 30%) - computed
PASS circle(at bottom 70% right 80%) serializes as circle(at 20% 30%) - computed
PASS circle(at bottom 70% right 80px) serializes as circle(at calc(100% - 80px) 30%) - computed
PASS circle(at bottom 70px left 50%) serializes as circle(at 50% calc(100% - 70px)) - computed
PASS circle(at bottom 70px right 50%) serializes as circle(at 50% calc(100% - 70px)) - computed
PASS circle(at bottom 70px right 80px) serializes as circle(at calc(100% - 80px) calc(100% - 70px)) - computed
PASS circle(at right 80% top 50%) serializes as circle(at 20% 50%) - computed
PASS circle(at right 80% top 50px) serializes as circle(at 20% 50px) - computed
PASS circle(at right 80% bottom 70%) serializes as circle(at 20% 30%) - computed
PASS circle(at right 80% bottom 70px) serializes as circle(at 20% calc(100% - 70px)) - computed
PASS circle(at right 80px top 50%) serializes as circle(at calc(100% - 80px) 50%) - computed
PASS circle(at right 80px bottom 70%) serializes as circle(at calc(100% - 80px) 30%) - computed
PASS circle(at right 80px bottom 70px) serializes as circle(at calc(100% - 80px) calc(100% - 70px)) - computed
